feat: entites 파일 분리 (Layers - Slices - Segments) 1. entites/comment/model - comment model 분리
